chore: rename dirs, chores
