+++ packages/ui/src/components/Notifications/SNotificationBody.vue
