+++ packages/ui/src/components/Notifications/api.ts
@@ -1,4 +1,4 @@
 import { InjectionKey } from 'vue'
-import { ToastsApi } from '../SToasts'
+import { ToastsApi } from '../Toasts'
 
 export const NOTIFICATIONS_API_KEY: InjectionKey<ToastsApi> = Symbol('NotificationsToastsAPI')
